requirements.txt
google.generativeai
emoji
openai
langchain
wikipedia-api
fpdf2
